% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bs_mean.R
\name{bs_mean}
\alias{bs_mean}
\title{Parametric bootstrapped means}
\usage{
bs_mean(z, n = 200, use.abs = FALSE)
}
\arguments{
\item{z}{Parameter estimates}

\item{n}{Number of bootstrap replications}

\item{use.abs}{Rank z by absolute value.}
}
\value{
A vector of mean estimates of the same length as \code{z}
}
\description{
This function implements the method described in Simon and Simon (2013)
}

