% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_sim.R
\name{cor_sim}
\alias{cor_sim}
\title{Calculate expected coverage of several alternative intervals for
 marginal coefficients in gene expression correlation example}
\usage{
cor_sim(Sigma, n, n.rep = 10, eS = NULL)
}
\arguments{
\item{Sigma}{Covariance matrix from which data will be generated.}

\item{n}{Number of samples}

\item{n.rep}{Number of replications}

\item{eS}{Optional eigen decomposition of Sigma}
}
\value{
A list with elements:
\describe{
\item{\code{COVERAGE}}{A 5 by p by n array giving the coverage at each simulation.
\code{COVERAGE[i, j, k]} is either 0 or 1 indicating weather the interval formed using
method i for rank j in simulation k covered its respective parameter}
\item{\code{WIDTH}}{ A 5 by p by n array giving the width of each interval in each simulation.}
 \item{\code{simnames}}{Name of each type of the 5 types of confidence intervals}
}
}
\description{
This function can be used to generate simulations in section 3.1
}

