% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg_sim.R
\name{linreg_sim}
\alias{linreg_sim}
\title{Calculate expected coverage of several alternative intervals for
 marginal coefficients in linear regression.}
\usage{
linreg_sim(X.pop, which.sample, n.rep = 10, which.X0 = c(), beta = c(0))
}
\arguments{
\item{X.pop}{a matrix of coefficients for an entire population.}

\item{which.sample}{Indices of individuals who are sampled}

\item{n.rep}{Number of simulations to run}

\item{which.X0}{Which coefficients contribute to the phenotype}

\item{beta}{Effect size (including an intercept). Phenotypes whill be generated as
\code{y = beta \%*\% X0} where
\code{X0 = cbind(rep(1, length(which.sample)), X.pop[which.sample, which.X0])}}
}
\value{
A list with elements:
\describe{
\item{\code{COVERAGE}}{A 5 by p by n array giving the coverage at each simulation.
\code{COVERAGE[i, j, k]} is either 0 or 1 indicating weather the interval formed using
method i for rank j in simulation k covered its respective parameter}
\item{\code{WIDTH}}{ A 5 by p by n array giving the width of each interval in each simulation.}
 \item{\code{simnames}}{Name of each type of the 5 types of confidence intervals}
}
}
\description{
This function can be used to generate simulations in section 3.1
}

