% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callMACS.R
\name{makemacsdata}
\alias{makemacsdata}
\title{Call MACS (Chen 2009) to generate genotype data}
\usage{
makemacsdata(N, length = 15000, t = 0.001, r = 0.001, seed = NULL,
  macs.loc = "~/macs/", tempdir = "./")
}
\value{
A matrix of genotypes
}

