% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par_bs_ci.R
\name{par_bs_ci}
\alias{par_bs_ci}
\title{Parametric bootstrapped confidence intervals to control RCC}
\usage{
par_bs_ci(beta, se = rep(1, length(beta)), rank.func = NULL, theta = beta,
  level = 0.9, n.rep = 1000, use.abs = TRUE, ...)
}
\arguments{
\item{beta}{Parameter estimates}

\item{se}{Estimated standard error of beta. Defaults to 1.}

\item{rank.func}{A Function that takes as first argument beta/se and returns a list with items order and rank.}

\item{theta}{Possibly shrunken estimates of E[beta]. Defaults to beta.}

\item{level}{Confidence level}

\item{n.rep}{Number of bootstrap replications}

\item{use.abs}{Base the rank on abs(beta) rather than beta}

\item{...}{Additional parameters to pass to rank.func}
}
\value{
A p by 2 matrix giving confidence intervals for each element of \code{beta}
}
\description{
This function implements algorithm 1 - simple
parametric bootstrap for normal estimates
}
