% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/par_bs_ci_isoreg.R
\name{par_bs_ci_sm}
\alias{par_bs_ci_sm}
\title{Parametric bootstrapped confidence intervals to control RCC}
\usage{
par_bs_ci_sm(z, theta = z, alpha = 0.1, n = 200, use.abs = FALSE)
}
\arguments{
\item{z}{Parameter estimates}

\item{theta}{Used to produce oracle intervals for simulations whien E[z] is known.}

\item{alpha}{Confidence level}

\item{n}{Number of bootstrap replications}

\item{use.abs}{Rank z by absolute value.}
}
\value{
A p by 2 matrix giving confidence intervals for each element of \code{z}
}
\description{
This function implements algorithm 1 - simple
parametric bootstrap for normal estimates
}

