% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par_bs_db_ci.R
\name{par_bs_db_ci}
\alias{par_bs_db_ci}
\title{Parametric bootstrapped confidence intervals to control RCC}
\usage{
par_bs_db_ci(z, alpha = 0.1, n = 2000, use.abs = FALSE, lazy = FALSE)
}
\arguments{
\item{z}{Parameter estimates}

\item{alpha}{Confidence level}

\item{n}{Number of bootstrap replications}

\item{use.abs}{Rank z by absolute value.}

\item{lazy}{Assume the same bias in bootstrap samples as estimated in the original sample.}
}
\value{
A p by 2 matrix giving confidence intervals for each element of \code{z}
}
\description{
This function implements algorithm 4 - simple
parametric bootstrap for normal estimates with debiased mean estimation.
Means are debiased using bootstrapping as described in Simon and Simon (2013)
}

