% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_example_sims.R
\name{plot_example_coverage}
\alias{plot_example_coverage}
\title{Plot coverage of example simulations}
\usage{
plot_example_coverage(R, main = "", proportion = 0.2, make.legend = FALSE,
  y.axis.off = FALSE)
}
\arguments{
\item{R}{A list produced by \code{example_sim}}

\item{main}{Title}

\item{proportion}{Proportion of statistics to output}

\item{make.legend}{Produce a legend?}

\item{y.axis.off}{Don't label the y axis?}
}
\value{
A ggplot object
}
\description{
Plot coverage of example simulations
}

