% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run1_linreg_sim.R
\name{run1_linreg_sim}
\alias{run1_linreg_sim}
\title{Run one linear regression simulation using builtin parameters.}
\usage{
run1_linreg_sim(beta.type, which.run, file.prefix, seed = NULL,
  parallel = FALSE)
}
\arguments{
\item{beta.type}{One of "none", "ten_equal", "ten_norm", or "mixture"}

\item{which.run}{An integer that will get included in the name of the output file.}

\item{file.prefix}{Simulations results will be saved to file.prefix_`beta.type`_n`which.run`.RData}

\item{seed}{Optional seed to set}
}
\value{
This function saves an object produced by `linreg_sims` to a file.
}
\description{
This function is a wrapper for linreg_sim and calls it using built-in data
`macs_params`. `macs_params` is a list containing the effect sizes and variables
used to create the simulations in section 3.1. This function writes data to an output
file and is intended to be run in parallell.
}

