% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonpar_bs_ci.R
\name{nonpar_bs_ci}
\alias{nonpar_bs_ci}
\title{Non-Parametric bootstrapped confidence intervals to control RCC}
\usage{
nonpar_bs_ci(data, analysis.func, rank.func = NULL, level = 0.1,
  n.rep = 1000, use.abs = TRUE, parallel = FALSE, ...)
}
\arguments{
\item{data}{An n by K matrix of data}

\item{analysis.func}{A function that performs the analysis. It must take exactly one argument (data) and return a
list or data frame including an item called 'estimate' and an item calle 'statistic'. 
These should both be vectors of length p, the number of parameters.}

\item{rank.func}{A Function that takes as first argument statistic as returned by analysis.func, as second argument use.abs
and returns a list with items order and rank.}

\item{level}{Confidence level}

\item{n.rep}{Number of bootstrap replications}

\item{use.abs}{Rank based on abs(z) rather than z}

\item{...}{Additional parameters to pass to rank.func}
}
\value{
A p by 2 matrix giving confidence intervals
}
\description{
Non-Parametric bootstrapped confidence intervals to control RCC
}

