% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_example_sims.R
\name{plot_example_coverage}
\alias{plot_example_coverage}
\title{Plot coverage of example simulations}
\usage{
plot_example_coverage(R, main = "", proportion = 0.2, y.axis.off = FALSE,
  legend.position = c(0.28, 0.4))
}
\arguments{
\item{R}{A list produced by \code{example_sim}}

\item{main}{Title}

\item{proportion}{Proportion of statistics to output}

\item{y.axis.off}{Don't label the y axis?}

\item{legend.position}{Where to put the legend.}
}
\value{
A ggplot object
}
\description{
Plot coverage of example simulations
}

